+++ lib/potassium/assets/config/clock.rb
@@ -1,0 +1,11 @@
+require 'clockwork'
+require_relative './boot'
+require_relative './environment'
+
+module Clockwork
+  # Example
+  #
+  # every(5.minutes, 'ScheduledJob') do
+  #   ScheduledJob.perform_later
+  # end
+end