+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Update [multi buildpack]([multi-buildpack-heroku]) buildback to the heroku version
+- Update ruby version on CI docker image
 
 [multi-buildpack-heroku]: http://github.com/ddollar/heroku-buildpack-multi