+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Configure Rspec to work with Paperclip
 - Update [multi buildpack]([multi-buildpack-heroku]) buildback to the heroku version
 - Update ruby version on CI docker image
 - Adds mailer support with AWS-SES and Sendgrid [#103]