+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Update ruby version on CI docker image
 - Adds mailer support with AWS-SES and Sendgrid [#103]
 
+FIX:
+- Run db tasks with bundle exec
+
 [multi-buildpack-heroku]: http://github.com/ddollar/heroku-buildpack-multi
 
 ## 2.2.0