+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 - Add draper gem to decorate models
 
+## 3.0.0
+
+Features:
+  - Rails 5.0.2
+  - `better_errors` and `binding_of_caller` gem in development
+  - Update `aws_sdk` to `~> 2.5`
+  - Update `paperclip` to `~> 5,0`
+
 ## 2.3.0
 
 Features: